Renamed the some scripts for Puppeteer
