+++ backstop_data/engine_scripts/puppet/clickAndHoverHelperAsync.js
