+++ backstop_data/engine_scripts/puppet/onBefore.js
@@ -1,4 +1,4 @@
 /* eslint-disable global-require */
 module.exports = async (page, scenario) => {
-  await require('./loadCookies')(page, scenario);
+  await require('./loadCookiesAsync')(page, scenario);
 };
