+++ backstop_data/engine_scripts/puppet/onReady.js
@@ -1,5 +1,5 @@
 /* eslint-disable global-require */
 module.exports = async (page, scenario) => {
   console.log(`SCENARIO > ${scenario.label}`);
-  await require('./clickAndHoverHelper')(page, scenario);
+  await require('./clickAndHoverHelperAsync')(page, scenario);
 };
